+++ src/pages/index.js
@@ -1,0 +1,6 @@
+import MainPage from "./MainPage";
+import LoginPage from "./LoginPage";
+import ProfilePage from "./ProfilePage";
+import ErrorPage from "./ErrorPage";
+
+export { MainPage, LoginPage, ProfilePage, ErrorPage };
