+++ src/ui/Footer.js
@@ -1,0 +1,7 @@
+const Footer = `
+  <footer class="bg-gray-200 p-4 text-center">
+    <p>&copy; 2024 항해플러스. All rights reserved.</p>
+  </footer>
+`;
+
+export default Footer;
