+++ src/ui/index.js
@@ -1,0 +1,5 @@
+import Header from "./Header";
+import Footer from "./Footer";
+import Nav from "./Nav";
+
+export { Header, Footer, Nav };
